+++ extension/src/Settings/SettingsMap.ts
@@ -63,6 +63,7 @@ export const settingsMap = new Map<string, keyof Settings>([
   ["NAB.EnableTelemetry", "enableTelemetry"],
   ["NAB.EnableTroubleshootingCommands", "enableTroubleshootingCommands"],
   ["NAB.EnableXliffCache", "enableXliffCache"],
+  ["NAB.PreferLockedTranslations", "preferLockedTranslations"],
 
   // Other extension's settings:
   ["al.packageCachePath", "packageCachePath"],